+++ contracts/test/AllStandardOpsTest.sol
@@ -20,7 +20,6 @@ contract AllStandardOpsTest is RainVM, VMState {
     State private _state;
 
     constructor(StateConfig memory config_) {
-
         /// These local opcode offsets are calculated as immutable but are
         /// really just compile time constants. They only depend on the
         /// imported libraries and contracts. These are calculated at
@@ -56,13 +55,13 @@ contract AllStandardOpsTest is RainVM, VMState {
         return _state;
     }
 
-    function fnPtrs() external view returns (uint[] memory) {
+    function fnPtrs() external view returns (uint256[] memory) {
         DispatchTable dispatchTable_ = AllStandardOps.dispatchTable();
         return dispatchTable_.fnPtrs();
     }
 
     /// Runs `eval` and stores full state.
-    function run(uint[] memory fnPtrs_) public {
+    function run(uint256[] memory fnPtrs_) public {
         DispatchTable dispatchTable_;
         dispatchTable_ = dispatchTable_.initialize(fnPtrs_);
         State memory state_ = _restore(vmStatePointer);
