+++ contracts/vm/ops/math/LogicOps.sol
@@ -45,7 +45,11 @@ library LogicOps {
     }
 
     // ISZERO
-    function isZero(uint, uint256 stackTopLocation_) internal pure returns (uint256) {
+    function isZero(uint256, uint256 stackTopLocation_)
+        internal
+        pure
+        returns (uint256)
+    {
         assembly {
             // The index doesn't change for iszero as there is
             // one input and output.
@@ -60,7 +64,7 @@ library LogicOps {
     // before EAGER_IF will select one of them. If both x_ and y_
     // are cheap (e.g. constant values) then this may also be the
     // simplest and cheapest way to select one of them.
-    function eagerIf(uint, uint256 stackTopLocation_)
+    function eagerIf(uint256, uint256 stackTopLocation_)
         internal
         pure
         returns (uint256)
@@ -80,7 +84,7 @@ library LogicOps {
         return stackTopLocation_;
     }
 
-    function equalTo(uint, uint256 stackTopLocation_)
+    function equalTo(uint256, uint256 stackTopLocation_)
         internal
         pure
         returns (uint256)
@@ -93,7 +97,7 @@ library LogicOps {
         return stackTopLocation_;
     }
 
-    function lessThan(uint, uint256 stackTopLocation_)
+    function lessThan(uint256, uint256 stackTopLocation_)
         internal
         pure
         returns (uint256)
@@ -106,7 +110,7 @@ library LogicOps {
         return stackTopLocation_;
     }
 
-    function greaterThan(uint, uint256 stackTopLocation_)
+    function greaterThan(uint256, uint256 stackTopLocation_)
         internal
         pure
         returns (uint256)
@@ -122,7 +126,7 @@ library LogicOps {
     // EVERY
     // EVERY is either the first item if every item is nonzero, else 0.
     // operand_ is the length of items to check.
-    function every(uint operand_, uint256 stackTopLocation_)
+    function every(uint256 operand_, uint256 stackTopLocation_)
         internal
         pure
         returns (uint256)
@@ -149,7 +153,7 @@ library LogicOps {
     // ANY
     // ANY is the first nonzero item, else 0.
     // operand_ id the length of items to check.
-    function any(uint operand_, uint256 stackTopLocation_)
+    function any(uint256 operand_, uint256 stackTopLocation_)
         internal
         view
         returns (uint256)
