+++ contracts/vm/ops/token/IERC721Ops.sol
@@ -19,47 +19,54 @@ library IERC721Ops {
         return -1;
     }
 
-            // Stack the return of `balanceOf`.
-    function balanceOf(uint, uint stackTopLocation_) internal view returns (uint) {
-                uint256 location_;
-                uint256 token_;
-                uint256 account_;
+    // Stack the return of `balanceOf`.
+    function balanceOf(uint256, uint256 stackTopLocation_)
+        internal
+        view
+        returns (uint256)
+    {
+        uint256 location_;
+        uint256 token_;
+        uint256 account_;
 
-                assembly {
-                    stackTopLocation_ := sub(stackTopLocation_, 0x20)
-                    location_ := sub(stackTopLocation_, 0x20)
-                    token_ := mload(location_)
-                    account_ := mload(stackTopLocation_)
-                }
-                uint256 balance_ = IERC721(address(uint160(token_))).balanceOf(
-                    address(uint160(account_))
-                );
+        assembly {
+            stackTopLocation_ := sub(stackTopLocation_, 0x20)
+            location_ := sub(stackTopLocation_, 0x20)
+            token_ := mload(location_)
+            account_ := mload(stackTopLocation_)
+        }
+        uint256 balance_ = IERC721(address(uint160(token_))).balanceOf(
+            address(uint160(account_))
+        );
 
-                assembly {
-                    mstore(location_, balance_)
-                }
-                return stackTopLocation_;
+        assembly {
+            mstore(location_, balance_)
+        }
+        return stackTopLocation_;
     }
 
-            // Stack the return of `ownerOf`.
-    function ownerOf(uint, uint stackTopLocation_) internal view returns (uint) {
-                        uint256 location_;
-                uint256 token_;
-                uint256 id_;
+    // Stack the return of `ownerOf`.
+    function ownerOf(uint256, uint256 stackTopLocation_)
+        internal
+        view
+        returns (uint256)
+    {
+        uint256 location_;
+        uint256 token_;
+        uint256 id_;
 
-                assembly {
-                    stackTopLocation_ := sub(stackTopLocation_, 0x20)
-                    location_ := sub(stackTopLocation_, 0x20)
-                    token_ := mload(location_)
-                    id_ := mload(stackTopLocation_)
-                }
-                uint256 owner_ = uint256(
-                    uint160(IERC721(address(uint160(token_))).ownerOf(id_))
-                );
-                assembly {
-                    mstore(location_, owner_)
-                }
-                return stackTopLocation_;
+        assembly {
+            stackTopLocation_ := sub(stackTopLocation_, 0x20)
+            location_ := sub(stackTopLocation_, 0x20)
+            token_ := mload(location_)
+            id_ := mload(stackTopLocation_)
+        }
+        uint256 owner_ = uint256(
+            uint160(IERC721(address(uint160(token_))).ownerOf(id_))
+        );
+        assembly {
+            mstore(location_, owner_)
+        }
+        return stackTopLocation_;
     }
-
 }
